+++ src/main/java/stroom/dashboard/expression/v1/FunctionFactory.java
@@ -69,6 +69,7 @@ public class FunctionFactory {
 
         add(Replace.class, Replace.NAME);
         add(Concat.class, Concat.NAME);
+        add(Link.class, Link.NAME);
 
         add(StringLength.class, StringLength.NAME);
         add(UpperCase.class, UpperCase.NAME);