+++ CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Added automatic module name for Java 9+ builds.
 
+## [v1.4.13] - 2018-12-06
+
+* Issue **#stroom#989** : Added `link` function.
+
+* Issue **#stroom#991** : Concat now supports a single parameter.
+
 ## [v1.4.12] - 2018-11-20
 
 * Issue **#22** : Added additional type checking functions `isBoolean()`, `isDouble()`, `isError()`, `isInteger()`, `isLong()`, `isNull()`, `isNumber()`, `isString()`, `isValue()`. Testing equality of null with `x=null()` is no longer valid and must be replaced with `isNull(x)`.